20220606033312 : embrace the fused multiply add for IEEE754-2008

+++ nvidia_cuda/dev_info/cuda_compile.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
 
-PATH=/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin:/usr/local/cuda-11.4/bin:/opt/schily/bin
+PATH=/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin:/opt/schily/bin
+
+if [ -d /usr/local/cuda-11.4 ]; then
+    CUDA_HOME=/usr/local/cuda-11.4
+    export CUDA_HOME
+    PATH=/usr/local/cuda-11.4/bin:$PATH
+    export PATH
+fi
 
 NVCC=`(command -v nvcc)` ; export NVCC
+if [ ! -x ${NVCC} ]; then
+    /usr/bin/printf "FAIL : the NVidia compiler not found\n"
+    return 1
+fi
 
 rm -f dev_info dev_info.o > /dev/null 2>&1
